Fixed a complex series of bugs in the Foreign (and other) fields use of __getattr__
The first is that self wasn't properly prefixed to instance.  The second was that
although the ForeignObject was being properly cloned, when we went to set up the keys
(in Gob.__init__) we were doing it on the original object (the field as it was on the
class).  The third was that clone (or base_clone, as I've renamed it) was depending on
being able to request certain methods which are not always present on an object:
'__getnewargs__', '__getstate__', and '__setstate__'.  Since this happened outside of
the domain of __init__ (not called for clone), then 'instance' was not yet set, so there
was the loop between value and instance that we were seeing recurse for a long time,
but not indefinitely.  Not indefinitely because eventually it threw a recursion depth
error, which was interpreted pretty much like a we didn't find it error, and so only
*then* would things move on, normally, as if there were no bug at all.  As this must
have happened multiple times on *every* gob creation, I think we should be able to
expect the performance to improve wildly now that this is fixed.  The fourth, and
final, bug was caused by self.value being set to None in Field.  So I just added a del
self.value *after* the call to super(...).__init__ on Foreign and now it works as
expected.