The query constraints sort of work now.
Actually, the constraints seemed to have already worked fine, but some of this other stuff is
a little fubar...

The main change to this is to parse the constraints properly, bugfixes from some code that
hadn't been hit before in query serialization (for caching queries), and making the
apiobj.Workspaces.list method at least a *little* sane.  Sort of.  A lot of this needs to be
gone over, since 98% correct code only counts in computational horseshoes.