Code audit round 1: gobpersist core (no specific back ends).
* Fixed size==0 bug on gobpersist.storage.Storable

Previously a zero size would have caused it to try to obtain the size from the transfer,
since there wasn't a distinction between zero and undefined.  Now there is.

* Fixed some erroneos % mongering on error messages in gobpersist.field

* gobpersist.field.StringField now is more strict about unicode<->binary translation.

If the encoding is 'binary', then we consider this to be impossible, but go ahead and try
with the defaults when requested.  Otherwise, translation always occurs both ways as part
of the verification.  This will probably break some stuff until upstream is changed as
well.

* Fixed mismatch between 'gte'/'lte' vs. 'ge'/'le' in gobpersist.session.

'ge'/'le' wins.

* Added delegation to file-like wrapper classes in gobpersist.storage

__getattr__ is now set for the file-like object wrappers, so these should be more usable
directly as files.

* Made the file-like object gobpersist.sesion.StorageEngine.upload_iter uses to call
StorageEngine.upload return as soon as there is any unblocked data.

As opposed to previously, when it would wait until it had read the requested amount.
This is just for the default implementation of upload_iter, a wrapper for upload.

* Factored SimpleThreadMappedPool out of gobpersist.backends.cache into
gobpersist.backends.pools.  Fixed potential leakage.

This simply didn't belong in the earlier location.  The leakage was fixed by calling a
cleanup function on every allocate.  For a large pool, with many, many threads, this
could be pretty unacceptable.  But it should be fine for most usages.  Note that this
change is not yet reflected in the gobpersist back ends.

* gobpersist.gobkvquerent now properly handles quantified identifiers of any arbitrary
depth.

This was a lot easier than I expected; just needed to define a recursive iterator.
gobkvquerent should now be a fully functional implementation of gobpersist, and any
arbitrary queries on it should work.