Moved non-generic gobpersist objects; made all orderings accept arbitrary identifiers.
This probably breaks the non-generic gobpersist stuff (all of which is for the file caching bit), as well as the stuff which uses it.

We may want to rework some of the "non-generic" stuff back into the main gobpersist library once we can make it generic.  That may mean making a generic fs storage engine.  In general, our file cache isn't generic since we can assume that files do not change.  If we can't easily break out a generic part, that's fine.  Shouldn't be a big deal.

As for the orderings, now you can go like this (...order={"asc": ("field", "subfield")}).  Unlike the generic quantifiers, this may have limited utility; but it's at least good for 1:1 mapping.