Fixes #29 by making sure the correct encoding is used between upload and add requests.
This was failing since the original file_handle returned from upload was in unicode, not the specified encoding specified so the followup add was requesting a different file handle than intended.  By making these explicit this problem goes away, but this still feels like a temporary solution.

We are now printing the string representation of potentially non-ascii data in the minifs print statements.  This might not be necessary once the logger is hooked up.

And we are corecing the filecahce location to a unicode string so we get valid results from os.listdir()