Got first round of consistency working -- cascading invalidation
This is less than ideal.  It does not properly lock.  And the use
of an integrity key to do the lookup is probably not ideal, nor is
it universal.  Nevertheless, it seems to work.

This fixes the problem where rootws wouldn't invalidate when a
top-level workspace was *removed*, but it does not fix the problem
where rootws wouldn't invalidate when a top-level workspace is
*added*.  For the latter, we should be creating or altering a
rootws element at a higher level (like in apiobj).