Rewrote the cache invalidation method to be properly recursive and atomic.
Previous version was sorta hacked.  Turns out this doesn't need
recursive locks, since we have to keep track of all the locks we
acquire outside of the recursive method anyway.  I'll probably back
those out and keep it simple.  gobpersist.exception.Deadlock is a
new exception class to handle an abort from arbitrary recursion
depth when a lock cannot be acquired.  However, this exception as
of now should not leak outside of the module, since we just force
acquire after a certain number of tries.